AudioFeatureExtractor/requirements.txt
absl-py==2.2.2
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
decorator==5.2.1
fastapi==0.115.12
grpcio==1.73.1
grpcio-tools==1.73.1
idna==3.10
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
msgpack==1.1.1
numba==0.61.2
numpy==2.1.3
packaging==25.0
pip==23.0.1
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
pycparser==2.22
python-multipart==0.0.20
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.2
soundfile==0.13.1
soxr==0.5.0.post1
threadpoolctl==3.6.0
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.2
